perf(lmstudio): keep thinking wrapper lightweight
